Move IfcUUID class to root of package
