--- src/compas_cadwork/utils/__init__.py
@@ -1,6 +1,0 @@
-from .ifc_uuid import IfcUUID
-
-
-__all__ = [
-    "IfcUUID",
-]
